runtime_monitor/requirements.txt
antlr4-python3-runtime==4.7.2
paho-mqtt==1.4.0
setuptools==41.0.1
pip==19.1.1